fix: multiple problems under Windows+MSVC+x86 (#4502)
Fixed a number of conversions from 64 to 32 bits, use the correct types
for file sizes vs. buffer sizes, etc. I disabled the `/WX` flag for this
platform because some Abseil (indirect) includes fail to build. I did
not fix all our warnings in tests and benchmarks.

+++ google/cloud/storage/benchmarks/throughput_experiment.h
@@ -27,8 +27,8 @@ namespace storage_benchmarks {
 struct ThroughputExperimentConfig {
   OpType op;
   std::int64_t object_size;
-  std::int64_t app_buffer_size;
-  std::uint64_t lib_buffer_size;
+  std::size_t app_buffer_size;
+  std::size_t lib_buffer_size;
   bool enable_crc32c;
   bool enable_md5;
 };
